Simplify local files by relying on _static
The previous implemented approach was copying files to the _images
directory, in a way to emulate Sphinx' own `.. image::` directive.
However this process can be a bit tricky.

We can instead simplify this implementation by assuming the favicon
files will always be placed in one of the directories listed in Sphinx
`html_static_path`.